refactor(application): 패키지 분리
작업 내용
- 가독성을 위해 패키지를 키워드로 분리
