+++ src/main/java/blackjack/domain/card/Hand.java
@@ -1,4 +1,4 @@
-package blackjack.domain.user;
+package blackjack.domain.card;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardValue;
