+++ src/main/java/blackjack/domain/user/Dealer.java
@@ -2,6 +2,7 @@ package blackjack.domain.user;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.game.GameRule;
+import blackjack.domain.value.Nickname;
 import blackjack.exception.ExceptionMessage;
 import java.util.List;
