+++ src/main/java/blackjack/domain/user/GameUser.java
@@ -1,6 +1,8 @@
 package blackjack.domain.user;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.card.Hand;
+import blackjack.domain.value.Nickname;
 import java.util.List;
 
 public class GameUser {
