+++ src/test/java/blackjack/domain/value/BettingAmountTest.java
@@ -1,8 +1,9 @@
-package blackjack.domain.user;
+package blackjack.domain.value;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.domain.value.BettingAmount;
 import blackjack.exception.ExceptionMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
